requirements.txt
addict==2.4.0
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
cityscapesScripts==2.2.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.1.1
crcmod==1.7
cycler==0.12.1
Cython==3.0.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.45.1
fqdn==1.5.1
fsspec==2023.12.2
humanfriendly==10.0
idna==3.6
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
# jupyterlab_server==2.25.2
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.0
model-index==0.1.11
mpmath==1.3.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
# opendatalab==0.0.10
# openmim==0.3.9
# openxlab==0.0.29
ordered-set==4.1.0
oss2==2.17.0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycryptodome==3.19.0
Pygments==2.17.2
pyparsing==3.1.1
#pyquaternion==0.9.9F
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
qudida==0.0.4
referencing==0.31.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.13.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
Send2Trash==1.8.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
terminado==0.18.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.65.2
traitlets==5.14.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.226
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
yapf==0.40.2
zipp==3.17.0
